energy-data-search/pyproject.toml
[project]
name = "energy-data-search"
version = "0.1.0"
description = "ChromaDB-powered search system for energy market documents (ISO rules, tariffs, proceedings)"
readme = "README.md"
authors = [
    { name = "Enrico Ladendorf", email = "enrico@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "click>=8.2.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.0",
    "unstructured>=0.18.13",
]

[project.scripts]
energy-search = "energy_data_search.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
